pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/epz_bot"]

[project]
name = "epz_bot_mwoe"
version = "4.1.0"
authors = [
    { name="Mara Wörle", email="mail@m-woerle.de" },
]
description = "A small Discord Bot to write into a Database"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
dependencies = [
    "mariadb==1.1.12",
    "discord.py==2.5.2",
    "typer==0.15.2",
    "pyyaml==6.0.2",
]

[project.scripts]
start = "epz_bot.cli:app"

[tool.mypy]
disable_error_code = ["import-untyped"]